tools/diary-converter/requirements.txt
google-generativeai>=0.3.0
pyyaml>=6.0
python-frontmatter>=1.0.0 